@@ -1,7 +1,7 @@
 {
-  "name": "react-emotion-starter",
+  "name": "mt4w",
   "version": "0.0.1",
-  "description": "a starter for projects using React & Emotion",
+  "description": "a website for Manual Therapy for Wellness",
   "main": "index.js",
   "scripts": {
     "build": "NODE_ENV=production webpack-cli --config ./webpack.config.js",
@@ -25,39 +25,42 @@
   },
   "homepage": "https://github.com/EvanMorrison/react-emotion-starter#readme",
   "dependencies": {
-    "@babel/polyfill": "^7.0.0",
+    "@babel/polyfill": "^7.4.4",
+    "@emotion/core": "^10.0.15",
+    "@emotion/styled": "^10.0.15",
     "@rmwc/button": "^3.0.9",
     "@rmwc/textfield": "^3.0.9",
-    "babel-plugin-emotion": "^9.2.11",
-    "emotion": "^9.2.12",
-    "emotion-theming": "^9.2.9",
-    "material-ui": "^0.20.2",
+    "emotion-theming": "^10.0.14",
+    "moment": "^2.24.0",
     "polished": "^2.3.0",
-    "prop-types": "^15.6.2",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "react-emotion": "^9.2.12",
+    "prop-types": "^15.7.2",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-hot-loader": "^4.3.12",
     "react-markdown": "^4.0.3",
-    "react-router-dom": "^4.3.1"
+    "react-router-dom": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/preset-env": "^7.1.0",
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.5.5",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.4",
-    "babel-preset-react-app": "^6.1.0",
+    "@emotion/babel-preset-css-prop": "^10.0.14",
+    "app-manifest-loader": "^2.4.0",
+    "babel-eslint": "^10.0.2",
+    "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^0.1.19",
-    "eslint": "^5.8.0",
+    "eslint": "^6.1.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-loader": "^2.1.1",
+    "eslint-plugin-emotion": "^10.0.14",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-node": "^8.0.0",
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-react-hooks": "^1.7.0",
     "eslint-plugin-standard": "^4.0.0",
     "file-loader": "^2.0.0",
     "html-webpack-plugin": "^3.2.0",
