@@ -2,16 +2,16 @@ import React, { Component } from 'react';
 import { PropTypes } from 'prop-types';
 import { Row } from '../AppStyles';
 import NavMenu from './NavMenu';
-import styled from 'react-emotion';
+import styled from '@emotion/styled';
 
 const DesktopNavWrapper = styled(Row)`
   flex: 1;
   justify-content: flex-end;
   align-items: center;
 
-  * {
-    color: ${props => props.position === 1 ? props.theme.contrastText : props.theme.primary} !important;
-  }
+  /* * {
+    color: ${props => props.position === 1 ? props.theme.contrastText : props.theme.primary};
+  } */
 
   @media (max-width: 1023px) {
     display: none;
@@ -20,10 +20,9 @@ const DesktopNavWrapper = styled(Row)`
 
 class DesktopNavComponent extends Component {
   render() {
-    return(
+    return (
       <DesktopNavWrapper position={this.props.position}>
-        <NavMenu position={this.props.position}
-          servicesOpen={false}/>
+        <NavMenu position={this.props.position}/>
       </DesktopNavWrapper>
     );
   }
