@@ -1,5 +1,5 @@
 import React from 'react';
-import styled from 'react-emotion';
+import styled from '@emotion/styled';
 import { PropTypes } from 'prop-types';
 import { rgba } from 'polished';
 import { Column } from '../../AppStyles';
