@@ -19,10 +19,10 @@ const isLocalhost = Boolean(
 );
 
 export default function register() {
-  if(process.env.NODE_ENV === 'production' && 'serviceWorker' in navigator) {
+  if (process.env.NODE_ENV === 'production' && 'serviceWorker' in navigator) {
     // The URL constructor is available in all browsers that support SW.
     const publicUrl = new URL(process.env.PUBLIC_URL, window.location);
-    if(publicUrl.origin !== window.location.origin) {
+    if (publicUrl.origin !== window.location.origin) {
       // Our service worker won't work if PUBLIC_URL is on a different origin
       // from what our page is served on. This might happen if a CDN is used to
       // serve assets; see https://github.com/facebookincubator/create-react-app/issues/2374
@@ -32,7 +32,7 @@ export default function register() {
     window.addEventListener('load', () => {
       const swUrl = `${process.env.PUBLIC_URL}/service-worker.js`;
 
-      if(isLocalhost) {
+      if (isLocalhost) {
         // This is running on localhost. Lets check if a service worker still exists or not.
         checkValidServiceWorker(swUrl);
       } else {
@@ -50,8 +50,8 @@ function registerValidSW(swUrl) {
       registration.onupdatefound = () => {
         const installingWorker = registration.installing;
         installingWorker.onstatechange = () => {
-          if(installingWorker.state === 'installed') {
-            if(navigator.serviceWorker.controller) {
+          if (installingWorker.state === 'installed') {
+            if (navigator.serviceWorker.controller) {
               // At this point, the old content will have been purged and
               // the fresh content will have been added to the cache.
               // It's the perfect time to display a "New content is
@@ -77,7 +77,7 @@ function checkValidServiceWorker(swUrl) {
   fetch(swUrl) // eslint-disable-line no-undef
     .then(response => {
       // Ensure service worker exists, and that we really are getting a JS file.
-      if(
+      if (
         response.status === 404 ||
         response.headers.get('content-type').indexOf('javascript') === -1
       ) {
@@ -100,7 +100,7 @@ function checkValidServiceWorker(swUrl) {
 }
 
 export function unregister() {
-  if('serviceWorker' in navigator) {
+  if ('serviceWorker' in navigator) {
     navigator.serviceWorker.ready.then(registration => {
       registration.unregister();
     });
